pyproject.toml
[tool.poetry]
name = "ranch-bot"
version = "0.1.0"
description = ""
authors = ["Erik Wahlberger <ewahlberger@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
aiohttp = "^3.8.1"
aiosignal = "^1.2.0"
async-timeout = "^4.0.2"
attrs = "^21.4.0"
charset-normalizer = "^2.0.10"
colorlog = "^6.6.0"
coverage = "^6.2"
frozenlist = "^1.2.0"
hikari = "^2.0.0-alpha.105"
hikari-lightbulb = "^2.1.3"
idna = "^3.3"
iniconfig = "^1.1.1"
multidict = "^5.2.0"
packaging = "^21.3"
pluggy = "^1.0.0"
pure25519 = "^0.0.1"
py = "^1.11.0"
pyparsing = "^3.0.6"
pytest-asyncio = "^0.17.0"
pytest-cov = "^3.0.0"
pytest = "^6.2.5"
python-dotenv = "^0.19.2"
tomli = "^2.0.0"
uvloop = "^0.16.0"
yarl = "^1.7.2"
colour = "^0.1.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
